+++ sdk/computelimit/azure-mgmt-computelimit/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Release History
 
+## 1.0.0 (2026-04-17)
+
+### Features Added
+
+  - Client `ComputeLimitMgmtClient` added operation group `features`
+  - Client `ComputeLimitMgmtClient` added operation group `vm_families`
+  - Added model `Feature`
+  - Added model `FeatureProperties`
+  - Added enum `FeatureState`
+  - Added model `OperationStatusResult`
+  - Added model `VmFamily`
+  - Added model `VmFamilyProperties`
+  - Added model `FeaturesOperations`
+  - Added model `VmFamiliesOperations`
+
 ## 1.0.0b1 (2025-11-14)
 
 ### Other Changes